Move appendProjectCode to a new project controller module.
